Fix some formatting and casting mistakes.

+++ src/rrdcached.c
@@ -89,7 +89,7 @@ static int value_list_to_string(char *buffer, int buffer_len,
 
     if (ds->ds[i].type == DS_TYPE_COUNTER) {
       status = snprintf(buffer + offset, buffer_len - offset, ":%" PRIu64,
-                        vl->values[i].counter);
+                        (uint64_t)vl->values[i].counter);
     } else if (ds->ds[i].type == DS_TYPE_GAUGE) {
       status = snprintf(buffer + offset, buffer_len - offset, ":%f",
                         vl->values[i].gauge);
